Added "India" Region
I have found that India has appeared as a region in a couple of NES games, so it report an error on the log:

"WARNING [TwinBee (India) (En) (Pirate)]: no recognizable regions found"

Those are the only two roms with "India" a as region (in all the No-Intro datasets of all systems :

https://datomatic.no-intro.org/index.php?page=show_record&s=45&n=2146

https://datomatic.no-intro.org/index.php?page=show_record&s=45&n=8157

Both are in english. There are no Hindi ("hi" as ISO language) language roms that are known of, at least belonging to the "India" region.